Move API console into web package
